feat: integrate shell, widen column to 59rem, delete dead layouts
