--- front_end/src/app/(main)/questions/[id]/components/question_layout/consumer_question_layout/consumer_tabs.tsx
@@ -1,24 +1,0 @@
-"use client";
-
-import { PropsWithChildren } from "react";
-
-import { Tabs } from "@/components/ui/tabs/index";
-
-import { useQuestionLayout } from "../question_layout_context";
-
-const ConsumerTabs: React.FC<PropsWithChildren> = ({ children }) => {
-  const { activeTab, setActiveTab } = useQuestionLayout();
-
-  return (
-    <Tabs
-      defaultValue="comments"
-      className="-mb-5"
-      value={activeTab}
-      onChange={setActiveTab}
-    >
-      {children}
-    </Tabs>
-  );
-};
-
-export default ConsumerTabs;
