--- front_end/src/app/(main)/questions/[id]/components/question_view/forecaster_question_view/index.tsx
@@ -1,46 +1,0 @@
-import { Fragment } from "react";
-
-import DetailedGroupCard from "@/components/detailed_question_card/detailed_group_card";
-import DetailedQuestionCard from "@/components/detailed_question_card/detailed_question_card";
-import ForecastMaker from "@/components/forecast_maker";
-import { PostStatus, PostWithForecasts } from "@/types/post";
-import {
-  isGroupOfQuestionsPost,
-  isQuestionPost,
-} from "@/utils/questions/helpers";
-
-import QuestionHeader from "./question_header";
-
-type Props = {
-  postData: PostWithForecasts;
-  preselectedGroupQuestionId?: number | undefined;
-};
-
-const ForecasterQuestionView: React.FC<Props> = ({
-  postData,
-  preselectedGroupQuestionId,
-}) => {
-  const isResolved = postData.status === PostStatus.RESOLVED;
-  const isGroup = isGroupOfQuestionsPost(postData);
-
-  return (
-    <Fragment>
-      <QuestionHeader post={postData} />
-      {isQuestionPost(postData) && (
-        <DetailedQuestionCard
-          post={postData}
-          keyFactors={postData.key_factors}
-        />
-      )}
-      {isGroup && (
-        <DetailedGroupCard
-          post={postData}
-          preselectedQuestionId={preselectedGroupQuestionId}
-        />
-      )}
-      {(!isResolved || isGroup) && <ForecastMaker post={postData} />}
-    </Fragment>
-  );
-};
-
-export default ForecasterQuestionView;
